+++ src/init.c
@@ -5,11 +5,15 @@
 extern SEXP lorentz_sup_c(SEXP, SEXP, SEXP, SEXP);
 extern SEXP find_peaks_c(SEXP);
 extern SEXP align_dp_c(SEXP, SEXP, SEXP);
+extern SEXP triangle_sup_c(SEXP, SEXP, SEXP, SEXP);
+extern SEXP rect_sup_c(SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef CallMethods[] = {
-    {"lorentz_sup_c", (DL_FUNC) &lorentz_sup_c, 4},
-    {"find_peaks_c",  (DL_FUNC) &find_peaks_c,  1},
-    {"align_dp_c",    (DL_FUNC) &align_dp_c,    3},
+    {"lorentz_sup_c",  (DL_FUNC) &lorentz_sup_c,  4},
+    {"find_peaks_c",   (DL_FUNC) &find_peaks_c,   1},
+    {"align_dp_c",     (DL_FUNC) &align_dp_c,     3},
+    {"triangle_sup_c", (DL_FUNC) &triangle_sup_c, 4},
+    {"rect_sup_c",     (DL_FUNC) &rect_sup_c,     4},
     {NULL, NULL, 0}
 };